requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mongo==0.2.0
pydantic==1.10.7
pymongo==4.3.3
python-multipart==0.0.6
pytz==2023.3
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
numpy==1.24.3
scipy==1.10.1
summa==1.2.0
joblib==1.2.0
nltk==3.8.1
regex==2023.3.23
tqdm==4.65.0
bcrypt==4.0.1
beanie==1.19.0
cffi==1.15.1
cryptography==40.0.2
email-validator==2.0.0.post2
fastapi-users-11.0.0
fastapi-users-db-beanie==2.0.0
lazy-model==0.0.5
makefun==1.15.1
motor==3.1.2
passlib==1.7.4
pycparser==2.21
pyjw==-2.6.0
toml==0.10.2
